refactor(config): extract provider-name leaf; break config<->profile cycle
